Return from the loop if a port name is found
The function devicePortName() search in the Registry the names of ports
by names of registry keys. At first looks for value of a key of "PortName",
and then of "PortNumber" key.

Thus, the first found value shall stop search and do not try to continue.

Tested on Windows 7/8 using Qt4 and then Qt5.

Change-Id: I98b00ff043a3b08476fec0a57b0d36ce65fc8d63
Reviewed-by: Sergey Belyashov <Sergey.Belyashov@gmail.com>