Add the synchronous loopback data verification test
Good reason to check of data correctness at transferring and receiving.
In this test are used two serial ports, connected in a null-modem mode:
http://en.wikipedia.org/wiki/Null_modem .

The sender port transfers data to the receiver port. After data are
received is carried out check of equivalence of the transferred and
received data.

Before run of testing it is necessary to set two variable environments
QTEST_SERIALPORT_SENDER and QTEST_SERIALPORT_RECEIVER to specify names
of used serial ports.

This test can reveal errors related with the internal data processing
and also errors of synchronous I/O inside of QSerialPort.

This test is placed in separate "benchmarks" category not to mix-up
with "manual" and "auto" categories, because this test is closer to
"benchmarks" by own functionality.

Tested on Windows 7/8 with the virtual com0com serial ports, using
Qt4 and then Qt5.

Change-Id: Ie6f87b50784bce211cf68c16cf75f79d12a8564f
Reviewed-by: Denis Shienkov <denis.shienkov@gmail.com>
Reviewed-by: Konstantin Ritt <ritt.ks@gmail.com>