Move the synchronous loopback test from benchmarks to auto tests
Previous commit 76f293bfb0fd7bab391f96ac822eacc40e4c5176 with adding of
this test to the benchmarks were hasty.

After all it makes sense to place it to auto tests to have opportunity to
run with others together.

Besides, this test is renamed since it is not a loopback, it is an I/O
test with the synchronous approach.

Change-Id: I043a6e0075561167bb29a59a384554ef98dbbd4a
Reviewed-by: Sergey Belyashov <Sergey.Belyashov@gmail.com>
Reviewed-by: Denis Shienkov <denis.shienkov@gmail.com>