Use internal read buffer of QIODevice
It is reasonable to use the internal read buffer of QIODevice instead of
additional QRingBuffer which currently are used. It will allow to reduce
an overhead when copying data as now there is not an excess chain with
QRingBuffer. Besides, the majority of I/O classes from QtCore (except
QProcess) also use the internal QIODevice buffer which are optimized for
this purpoze.

Tested on Windows 8 and Linux with the virtual com0com and on-board
serial ports.

Change-Id: I6c3d4b84593940ccfef8c15c9fbef16bddcd002f
Reviewed-by: Sergey Belyashov <Sergey.Belyashov@gmail.com>