Allow to use custom devices paths
QSP incorrectly transforms non-standard device names to their paths
and vice-versa (for example, "/home/ttyS0", "//./COM1", and so on).

Now this problem is solved:

* The transformation code is moved to QSPP.

* Added autotests auto/qserialportinfoprivate to testing of
conversion algorithm. These tests are private and can be activated
with building of QtSerialPort with:

qmake "QT_CONFIG+=private_tests warnings_are_errors" \
		DEFINES+=QT_BUILD_INTERNAL

Tested on Windows 8, Linux, OSX with auto-tests, with on-board
and virtual serial ports.

Task-number: QTBUG-38639
Change-Id: I43757a7f1390f53ed2b1d70de59c6bfb71892a59
Reviewed-by: Denis Shienkov <denis.shienkov@gmail.com>