Do not use internal pre-cached RTS and DTR variables
Some devices do not support these signals, and returns an errors when attempt
to query it. But the QSP::open() always requests a status of these lines.
In this case the QSP::open() will return true, but also will setup an error
code (for example, QSP::UnsupportedOperationError for socat's devices in
Linux). That is unexpected behavior for the user.

It is reasonable do not query a state of these signals when opening.
Also we do not need to use of appropriate internal variables, and we can
remove it.

Change-Id: I277110014c20c1f2c3fdaed20c4582d1356897b9
Reviewed-by: Denis Shienkov <denis.shienkov@gmail.com>