Share the clearReceiver() function and for the *nix tests
The waitForReadyReadWithTimeout() test is failed when using the socat
utility. The reason is that the receiver's side receive data from the
previous waitForBytesWritten() test, even the receiver was closed and
re-opened again in the new test. We need suppress this behavior using
the clearReceiver() function when starting of
waitForReadyReadWithTimeout() test.

Change-Id: I6735aa9c9744ec1732f9bd4f701068a540ae2316
Reviewed-by: Laszlo Papp <lpapp@kde.org>
Reviewed-by: Sergey Belyashov <Sergey.Belyashov@gmail.com>