Cleanup an error string when QSP::clearError() is called
QSP::clearError() calls setError() with an empty error string.
But setError() anyway calls the qt_error_string() that sets an
error string from an system error which occurred before, that
is garbage in this case.

Thus, we need to call qt_error_string() only if the input
QSP::SerialPortError is not NoError.

Change-Id: I0260159d74579de47bcf1c7b2f4e747be6817269
Reviewed-by: Sergey Belyashov <Sergey.Belyashov@gmail.com>
Reviewed-by: Denis Shienkov <denis.shienkov@gmail.com>