Improve the processing of errors
Sometimes the error string would contain a wrong description which
did not correspond to the system error code. The reason was
qt_error_string() being called too late, when the system error
might have already been overwritten.

The error processing is now in QSPP::getSystemError(), which returns both
the error code and the error description as soon as possible.

* Now the QSPP::getSystemError() returns the new class
  QSerialPortErrorInfo which contains all necessary fields.

* The new method QSPP::setError() which accepts the
  QSerialPortErrorInfo as input parameter is used.

* The old private method QSP::setError() is removed, because
  it is not used anywhere.

Change-Id: Ia7e4d617b863e2131175c52812cdf426ed963795
Reviewed-by: Sergey Belyashov <Sergey.Belyashov@gmail.com>
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@theqtcompany.com>
Reviewed-by: Denis Shienkov <denis.shienkov@gmail.com>