Obtain detail information in FreeBSD
Implementation based on the sysctl(3) system call,
in which undocumented features are used.

This implementation tested only with the USB devices,
because there is no opportunity to check it as well with
an other stuff.

Also still it is impossible to extract the description
and the manufacturer properties from the source, because
they are merged inside of one string.

Tested with the USB devices (FTDI, Prolific, ZTE, Samsung)
using FreeBSD v10.

(cherry-picked from 3e6308409dd679389e6231745e4d919d68d6487c)
Change-Id: I0091e4db70bfdfd4da199dd9d89dc78cf8f632b0
Reviewed-by: Denis Shienkov <denis.shienkov@gmail.com>