Use the termios v2 to custom baud rate switching
For a long time the Linux kernel has support new version of
the termios structure (V2). To setup the user speed it is
necessary to do following steps:

* Query current content of termios2 by calling the ioctl() with
  the TCGETS2 flag.

* Add to the c_cflag field the BOTHER flag.

* Set to the c_ispeed/c_ospeed fields an actual values of custom speeds.

* Write new content of termios2 back by calling the ioctl() with the
  TCSETS2 flag.

This new method much simpler and is more transparent in implementation
than changes of a custom divisor. It is preferable and will be used
by default for all cases. If for some reason a current Linux kernel
doesn't support the termios2, then will be falling back to the old
method with changing of a custom divisor.

Tested with the on-board and the USB (PL2303) serial ports.

(cherry-picked from eb0535c9701202d5f74fe1fc9b3b094bbecb8043)
Task-number: QTBUG-48094
Change-Id: I49a5389b089980b616b4e2ff815ce0b579752d0e
Reviewed-by: Denis Shienkov <denis.shienkov@gmail.com>