Allow to use the QSP::IgnorePolicy only
It is impossible to fix the dataErrorPolicy property that it worked at
any platforms (this property does not work at all), besides, this
property was deprecated. Therefore we can use always only the
IgnorePolicy value for all cases, that will not be worse than it is.

It gives a chance to get rid of internal code, related to dataErrorPolicy
handling in favor to fixing of the Mark/Space parity handling, and the
Parity/Frame/BreackCondition errors handling.

(cherry-picked from 918b22aef140d293721f0d81ea8988c1d35db582)
Change-Id: Ib0c5d1396819aabe459002c5ca5cca5840052eff
Reviewed-by: Denis Shienkov <denis.shienkov@gmail.com>