Cleanup remainders of code relating to data error policy handling
It is impossible to implement all these features on all platforms,
and particular drivers may not support them, too.

Consequently, the user should handle such errors themselves by applying
platform-specific ioctls on the device descriptor and/or parsing the
stream's byte-stuffing.

This commit also deprecates ParityError, FramingError, and
BreakConditionError.

Tested on Windows and Linux with the virtual and the
USB serial ports.

Change-Id: I4ffc2f067787bc304a83326acb2a2421b428f986
Reviewed-by: Sergey Belyashov <Sergey.Belyashov@gmail.com>
Reviewed-by: Denis Shienkov <denis.shienkov@gmail.com>