Fix build on Android x86
In the Android x86 NDK, the termios2 struct is already defined.
On ARM, it isn't defined until android-21, which is not what we
build against by default.

We should probably look for a more robust solution to this, to
support platforms where this is defined, but for now the main
issue is fixing the build.

Change-Id: I6084a90f664f9f778e6d1689106a3fb9a7ad0b2f
Reviewed-by: Alex Blasche <alexander.blasche@theqtcompany.com>