Forbid the RTS changing when the HardwareFlowControl was set
It is an error when the RTS changes manually when the
HardwareFlowControl is set. In this case the driver operates with the
RTS automatically, so manual control is not allowed. But some drivers
ignore this requirement and do not return errors in attempt of manual
RTS control. Thus, we have to forbid change of RTS and to return the
UnsupportedOperationError error.

Change-Id: I80bc87d51d9c653ed15ec310fca654243821b1e7
Reviewed-by: Sergey Belyashov <Sergey.Belyashov@gmail.com>
Reviewed-by: Denis Shienkov <denis.shienkov@gmail.com>